runtime: Add support for GDB JIT Interface
This is an initial attempt to support GDB JIT Interface in Dart VM.
More info about debugger interface itself can be found at [0].

To use this, there is an option added to dart. An example:

  $ dart --gdb-jit-interface hello.dart

We do not use DWARF for debug info production. We picked a simpler
custom debug info (proposed by feature [1]) for this initial step.

There is an LLDB plugin that understands this custom debug info [2].

[0] https://sourceware.org/gdb/current/onlinedocs/gdb.html/JIT-Interface.html
[1] https://sourceware.org/gdb/current/onlinedocs/gdb.html/Custom-Debug-Info.html#Custom-Debug-Info
[2] https://github.com/djtodoro/dart-lldb